requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.3
pywin32==305
pywinauto==0.6.8
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.2
